[*] Remove `concurrency` from `ci.yml` (#14)
<!-- Thanks for contributing to Artemis! Before you submit your pull
request, please make sure to check the following boxes by putting an x
in the [ ] (don't: [x ], [ x], do: [x]) -->
<!-- If your pull request is not ready for review yet, create a draft
pull request! -->

### Checklist
#### General
<!-- You only need to choose one of the first two check items:
Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and
you may remove the first checkmark. If you are unsure, please test on
the test servers. -->
- [ ] I locally pass **all** tests or calculated manually the correct
solutions on paper.
- [ ] I chose a title conforming to the [naming conventions for pull
requests](https://github.com/JohannesStoehr/gad23-tests/blob/main/CONTRIBUTING.md).
- [ ] The newly added tests do not reveal the solutions dynamically
calculating the solutions as expected in the exercise.

### Description
<!-- Describe your changes shortly -->

I was the person who added the `concurrency` feature for the PGdP CI but
I think the concurrency feature is for this kind of workflow completely
overkill because the workflow runs for a few seconds. It would be
something else if we had a workflow that runs for +10 minutes.
Additionally, because this repository is open source, we don't pay for
any GitHub Actions minutes (completely free). We only have the limit of
20 simultaneously Linux jobs but we will never reach this.

### Review Progress
<!-- Each Pull Request should be reviewed by at least one other student.
-->
<!-- The code and the functionality (= running the tests locally) need
to be reviewed -->
<!-- The reviewer or author check the following boxes. -->

#### Review
- [ ] Review 1
- [ ] Optional Review 2