[W02] Add test for the case when to and from of the returned interval are the same indices (#20)
### Checklist
#### General
<!-- You only need to choose one of the first two check items:
Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and
you may remove the first checkmark. If you are unsure, please test on
the test servers. -->
- [x] I locally pass **all** tests or calculated manually the correct
solutions on paper.
- [x] I chose a title conforming to the [naming conventions for pull
requests](https://github.com/JohannesStoehr/gad23-tests/blob/main/CONTRIBUTING.md).
- [x] The newly added tests do not reveal the solutions dynamically
calculating the solutions as expected in the exercise.

### Description

I passed all local but not all Artemis tests. This was because I forgot
the case where `from` and `to` of the returned interval are the same
number. The added test catches this edge case.

### Review Progress
<!-- Each Pull Request should be reviewed by at least one other student.
-->
<!-- The code and the functionality (= running the tests locally) need
to be reviewed -->
<!-- The reviewer or author check the following boxes. -->

#### Review
- [ ] Review 1
- [ ] Optional Review 2