[*] Contributing: replace PGDP keywords with GAD (#22)
<!-- Thanks for contributing to Artemis! Before you submit your pull
request, please make sure to check the following boxes by putting an x
in the [ ] (don't: [x ], [ x], do: [x]) -->
<!-- If your pull request is not ready for review yet, create a draft
pull request! -->

### Checklist
#### General
<!-- You only need to choose one of the first two check items:
Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and
you may remove the first checkmark. If you are unsure, please test on
the test servers. -->
- [x] I locally pass **all** tests or calculated manually the correct
solutions on paper.
- [x] I chose a title conforming to the [naming conventions for pull
requests](https://github.com/JohannesStoehr/gad23-tests/blob/main/CONTRIBUTING.md).
- [x] The newly added tests do not reveal the solutions dynamically
calculating the solutions as expected in the exercise.

### Description
<!-- Describe your changes shortly -->
The Contributions-Guideline contained the keyword "pgdp" instead of
"gad" in the section "Follow the package structure of the
implementation".

### Review Progress
<!-- Each Pull Request should be reviewed by at least one other student.
-->
<!-- The code and the functionality (= running the tests locally) need
to be reviewed -->
<!-- The reviewer or author check the following boxes. -->

#### Review
- [x] Review 1
- [x] Optional Review 2

Co-authored-by: danielpavlov <ge84nun@mytum.de>