[*] Rewrite PR Template for more informative general bullet points (#11)
<!-- Thanks for contributing to Artemis! Before you submit your pull
request, please make sure to check the following boxes by putting an x
in the [ ] (don't: [x ], [ x], do: [x]) -->
<!-- If your pull request is not ready for review yet, create a draft
pull request! -->

### Checklist
#### General
<!-- You only need to choose one of the first two check items:
Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and
you may remove the first checkmark. If you are unsure, please test on
the test servers. -->
- ~~I locally pass **all** tests or calculated manually the correct
solutions on paper.~~
- [x] I chose a title conforming to the [naming conventions for pull
requests](https://github.com/JohannesStoehr/gad23-tests/blob/main/CONTRIBUTING.md).
- [x] The newly added tests do not reveal the solutions dynamically
calculating the solutions as expected in the exercise.

### Description

The current checkboxes in general do not specify enough information
about the correctness of the tests, hence the split of 'I pass' to 'I
pass local tests', 'I pass artemis tests' and 'I guarantee the
correctness of tests'

It is possible to pass all local tests, but not on Artemis, also
otherwise.

### Review Progress
<!-- Each Pull Request should be reviewed by at least one other student.
-->
<!-- The code and the functionality (= running the tests locally) need
to be reviewed -->
<!-- The reviewer or author check the following boxes. -->

#### Review
- [ ] Review 1
- [ ] Optional Review 2