[W07] Adding last year tests and tests in RadixSortTest and BinaryRadixSortTest (#35)
<!-- Thanks for contributing to Artemis! Before you submit your pull
request, please make sure to check the following boxes by putting an x
in the [ ] (don't: [x ], [ x], do: [x]) -->
<!-- If your pull request is not ready for review yet, create a draft
pull request! -->

### Checklist
#### General
<!-- You only need to choose one of the first two check items:
Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and
you may remove the first checkmark. If you are unsure, please test on
the test servers. -->
- [x] I confirm that my tests, if any, are correct
    - [x] Optional: I pass **all** artemis tests.
    - [x] Optional: I pass **all** tests, that are added in this PR.
- [x] Optional: I calculated manually the correct solutions on paper.
- [x] I chose a title conforming to the [naming conventions for pull
requests](https://github.com/JohannesStoehr/gad23-tests/blob/main/CONTRIBUTING.md).
- [x] The newly added tests do not reveal the solutions dynamically
calculating the solutions as expected in the exercise.
- [x] Other students approved this Pull Request.

### Description
<!-- Describe your changes shortly -->
- `RadixSortTest`: I added two Edge-Casas for `provideKey` and in
`provideKeyInvalidIndex` invalid Indices
- `BinaryRadixSortTest`: I added `sortOnlyNegativTest` which test if
negativ numbers are sorted correcly only works if `lastSort` is
implemented, `sortOnlyPositivTest` which test if positiv numbers are
sorted correcly, `sortUsualTest` contains positiv and negativ Numbers,
also `lastSort` must be implemented, that this test works and
`sortBigTest` is the same as `sortUsualTest` but there are more numbers
which should be sorted. This test could fail on some devices but is also
repeated in the code.