[W08] Added tests from gad 2021 and tests in BinomialTreeNodeTest (#39)
<!-- Thanks for contributing to Artemis! Before you submit your pull
request, please make sure to check the following boxes by putting an x
in the [ ] (don't: [x ], [ x], do: [x]) -->
<!-- If your pull request is not ready for review yet, create a draft
pull request! -->

### Checklist
#### General
<!-- You only need to choose one of the first two check items:
Generally, test on the test servers. -->
<!-- If it's only a small change, testing it locally is acceptable and
you may remove the first checkmark. If you are unsure, please test on
the test servers. -->
- [x] I confirm that my tests, if any, are correct
    - [x] Optional: I pass **all** artemis tests.
    - [x] Optional: I pass **all** tests, that are added in this PR.
- [x] Optional: I calculated manually the correct solutions on paper.
- [x] I chose a title conforming to the [naming conventions for pull
requests](https://github.com/JohannesStoehr/gad23-tests/blob/main/CONTRIBUTING.md).
- [x] The newly added tests do not reveal the solutions dynamically
calculating the solutions as expected in the exercise.
- [x] Other students approved this Pull Request.

### Description
<!-- Describe your changes shortly -->
- Most of the tests are from the GAD-Course of 2021
- I only added tests in `BinomialTreeNodeTest` in form of
`minSimpleTest` and `rankSimpleTest` thease tests test if the right
return values is returned
- There are more tests from 2021 but due to the fact that I don't pass
them, I will not attach them to this PR.
- The class `StudentResultTest` is the same as `StudentResult` in the
project but with no `System.out.` which leds to a faster runtime
- The class `TreeCounterResult` is like `StudentResultTest` but is
needed for the `testAmountOFTrees` test in `BinomialHeapTest` and is
also from 2021
- The logging tests in class `LoggingTest` are comparing my output with
yours, to pass this tests you have to use the logging method as shown in
the `LoggingTest` file.
- The two files `insertLogging` and `insertDeleteLogging` provide the
expected logging
- In the class `BinomialHeapTest` are tree performance tests how they
work is written down in the file as a commant